requirements.txt
tensorflow-cpu==2.21.0
pillow==12.2.0
